requirements.txt
fastapi
uvicorn[standard]
pytest
pytest-asyncio
mongomock
mongomock-motor
motor
pydantic
cryptography
httpx

# LLM API clients
openai
anthropic
google-generativeai
litellm>=1.35.0

# Context management
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# Utilities
tiktoken
pyyaml
python-dotenv
tqdm
tenacity

# Streamlit for dashboard (Phase 4)
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.0.0
pandas>=2.0.0
numpy>=1.24.0

# Development tools
black
flake8
mypy
